+++ backend/canisters/escrow/api/src/updates/mod.rs
@@ -1,3 +1,3 @@
-pub mod cancel_offer;
-pub mod create_offer;
+pub mod cancel_swap;
+pub mod create_swap;
 pub mod notify_deposit;