+++ backend/canisters/escrow/c2c_client/src/lib.rs
@@ -4,6 +4,6 @@ use escrow_canister::*;
 // Queries
 
 // Updates
-generate_c2c_call!(create_offer);
-generate_c2c_call!(cancel_offer);
+generate_c2c_call!(create_swap);
+generate_c2c_call!(cancel_swap);
 generate_c2c_call!(notify_deposit);