+++ backend/canisters/user/impl/src/timer_job_types.rs
@@ -1,7 +1,5 @@
-use crate::model::p2p_swaps::P2PSwapOfferStatus;
 use crate::model::token_swaps::TokenSwap;
 use crate::updates::send_message::send_to_recipients_canister;
-use crate::updates::send_message_with_transfer::update_p2p_swap_status;
 use crate::updates::swap_tokens::process_token_swap;
 use crate::{mutate_state, openchat_bot, read_state};
 use canister_timer_jobs::Job;
@@ -66,13 +64,13 @@ pub struct ProcessTokenSwapJob {
 
 #[derive(Serialize, Deserialize, Clone)]
 pub struct NotifyEscrowCanisterOfDepositJob {
-    pub offer_id: u32,
+    pub swap_id: u32,
     pub attempt: u32,
 }
 
 impl NotifyEscrowCanisterOfDepositJob {
-    pub fn run(offer_id: u32) {
-        let job = NotifyEscrowCanisterOfDepositJob { offer_id, attempt: 0 };
+    pub fn run(swap_id: u32) {
+        let job = NotifyEscrowCanisterOfDepositJob { swap_id, attempt: 0 };
         job.execute();
     }
 }
@@ -81,7 +79,7 @@ impl NotifyEscrowCanisterOfDepositJob {
 pub struct SendMessageToGroupJob {
     pub chat_id: ChatId,
     pub args: group_canister::c2c_send_message::Args,
-    pub p2p_offer_id: Option<u32>,
+    pub p2p_swap_id: Option<u32>,
     pub attempt: u32,
 }
 
@@ -89,7 +87,7 @@ pub struct SendMessageToGroupJob {
 pub struct SendMessageToChannelJob {
     pub community_id: CommunityId,
     pub args: community_canister::c2c_send_message::Args,
-    pub p2p_offer_id: Option<u32>,
+    pub p2p_swap_id: Option<u32>,
     pub attempt: u32,
 }
 
@@ -210,7 +208,7 @@ impl Job for NotifyEscrowCanisterOfDepositJob {
             match escrow_canister_c2c_client::notify_deposit(
                 escrow_canister_id,
                 &escrow_canister::notify_deposit::Args {
-                    offer_id: self.offer_id,
+                    swap_id: self.swap_id,
                     user_id: None,
                 },
             )
@@ -222,7 +220,7 @@ impl Job for NotifyEscrowCanisterOfDepositJob {
                         let now = state.env.now();
                         state.data.timer_jobs.enqueue_job(
                             TimerJob::NotifyEscrowCanisterOfDeposit(Box::new(NotifyEscrowCanisterOfDepositJob {
-                                offer_id: self.offer_id,
+                                swap_id: self.swap_id,
                                 attempt: self.attempt + 1,
                             })),
                             now + 10 * SECOND_IN_MS,
@@ -240,11 +238,7 @@ impl Job for SendMessageToGroupJob {
     fn execute(self) {
         ic_cdk::spawn(async move {
             match group_canister_c2c_client::c2c_send_message(self.chat_id.into(), &self.args).await {
-                Ok(group_canister::c2c_send_message::Response::Success(_)) => {
-                    if let Some(id) = self.p2p_offer_id {
-                        update_p2p_swap_status(id, P2PSwapOfferStatus::Open);
-                    }
-                }
+                Ok(group_canister::c2c_send_message::Response::Success(_)) => {}
                 Err(_) if self.attempt < 20 => {
                     mutate_state(|state| {
                         let now = state.env.now();
@@ -252,7 +246,7 @@ impl Job for SendMessageToGroupJob {
                             TimerJob::SendMessageToGroup(Box::new(SendMessageToGroupJob {
                                 chat_id: self.chat_id,
                                 args: self.args,
-                                p2p_offer_id: self.p2p_offer_id,
+                                p2p_swap_id: self.p2p_swap_id,
                                 attempt: self.attempt + 1,
                             })),
                             now + 10 * SECOND_IN_MS,
@@ -270,11 +264,7 @@ impl Job for SendMessageToChannelJob {
     fn execute(self) {
         ic_cdk::spawn(async move {
             match community_canister_c2c_client::c2c_send_message(self.community_id.into(), &self.args).await {
-                Ok(community_canister::c2c_send_message::Response::Success(_)) => {
-                    if let Some(id) = self.p2p_offer_id {
-                        update_p2p_swap_status(id, P2PSwapOfferStatus::Open);
-                    }
-                }
+                Ok(community_canister::c2c_send_message::Response::Success(_)) => {}
                 Err(_) if self.attempt < 20 => {
                     mutate_state(|state| {
                         let now = state.env.now();
@@ -282,7 +272,7 @@ impl Job for SendMessageToChannelJob {
                             TimerJob::SendMessageToChannel(Box::new(SendMessageToChannelJob {
                                 community_id: self.community_id,
                                 args: self.args,
-                                p2p_offer_id: self.p2p_offer_id,
+                                p2p_swap_id: self.p2p_swap_id,
                                 attempt: self.attempt + 1,
                             })),
                             now + 10 * SECOND_IN_MS,