+++ backend/canisters/user_index/api/src/queries/reported_messages.rs
@@ -1,16 +1,18 @@
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
+use types::UserId;
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub struct Args {
-    pub offer_id: u32,
+    pub user_id: Option<UserId>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
 pub enum Response {
-    Success,
-    OfferAlreadyAccepted,
-    OfferExpired,
-    OfferNotFound,
-    NotAuthorized,
+    Success(SuccessResult),
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct SuccessResult {
+    pub json: String,
 }