+++ backend/integration_tests/src/client/user.rs
@@ -9,6 +9,7 @@ generate_query_call!(saved_crypto_accounts);
 generate_query_call!(updates);
 
 // Updates
+generate_update_call!(accept_p2p_swap);
 generate_update_call!(add_reaction);
 generate_update_call!(block_user);
 generate_update_call!(cancel_message_reminder);