+++ backend/libraries/types/src/p2p_swaps.rs
@@ -1,40 +1,61 @@
-use crate::icrc1::CompletedCryptoTransaction;
-use crate::{P2PSwapStatus, TimestampMillis, TransactionId, UserId};
+use crate::{P2PSwapContent, TimestampMillis, TransactionId, UserId};
 use candid::CandidType;
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Debug)]
-pub enum OfferStatus {
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub enum P2PSwapStatus {
     Open,
-    Cancelled(Box<OfferStatusCancelled>),
-    Expired,
-    Accepted(Box<OfferStatusAccepted>),
-    Completed(Box<OfferStatusCompleted>),
+    Cancelled(P2PSwapCancelled),
+    Expired(P2PSwapExpired),
+    Reserved(P2PSwapReserved),
+    Accepted(P2PSwapAccepted),
+    Completed(P2PSwapCompleted),
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OfferStatusCancelled {
-    pub cancelled_at: TimestampMillis,
+#[allow(clippy::large_enum_variant)]
+pub enum ReserveP2PSwapResult {
+    Success(ReserveP2PSwapSuccess),
+    Failure(P2PSwapStatus),
+    SwapNotFound,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OfferStatusAccepted {
+pub struct ReserveP2PSwapSuccess {
+    pub content: P2PSwapContent,
+    pub created: TimestampMillis,
+    pub created_by: UserId,
+}
+
+#[allow(clippy::large_enum_variant)]
+pub enum AcceptP2PSwapResult {
+    Success(P2PSwapAccepted),
+    Failure(P2PSwapStatus),
+    SwapNotFound,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct P2PSwapCancelled {
+    pub token0_txn_out: Option<TransactionId>,
+}
+
+pub type P2PSwapExpired = P2PSwapCancelled;
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct P2PSwapReserved {
+    pub reserved_by: UserId,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct P2PSwapAccepted {
     pub accepted_by: UserId,
-    pub accepted_at: TimestampMillis,
+    pub token1_txn_in: TransactionId,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OfferStatusCompleted {
+#[derive(CandidType, Serialize, Deserialize, Clone, Debug)]
+pub struct P2PSwapCompleted {
     pub accepted_by: UserId,
-    pub accepted_at: TimestampMillis,
-    pub token0_transfer_out: CompletedCryptoTransaction,
-    pub token1_transfer_out: CompletedCryptoTransaction,
-}
-
-#[derive(Serialize, Deserialize, Debug)]
-pub struct OfferStatusChange {
-    pub offer_id: u32,
-    pub status: OfferStatus,
+    pub token1_txn_in: TransactionId,
+    pub token0_txn_out: TransactionId,
+    pub token1_txn_out: TransactionId,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -47,8 +68,8 @@ pub enum AcceptSwapStatusError {
     AlreadyReserved(AcceptSwapAlreadyReserved),
     AlreadyAccepted(AcceptSwapAlreadyAccepted),
     AlreadyCompleted(AcceptSwapAlreadyCompleted),
-    OfferExpired(AcceptSwapOfferExpired),
-    OfferCancelled(AcceptSwapOfferCancelled),
+    SwapExpired(AcceptSwapSwapExpired),
+    SwapCancelled(AcceptSwapSwapCancelled),
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
@@ -71,12 +92,12 @@ pub struct AcceptSwapAlreadyCompleted {
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
-pub struct AcceptSwapOfferExpired {
+pub struct AcceptSwapSwapExpired {
     pub token0_txn_out: Option<TransactionId>,
 }
 
 #[derive(CandidType, Serialize, Deserialize, Debug)]
-pub struct AcceptSwapOfferCancelled {
+pub struct AcceptSwapSwapCancelled {
     pub token0_txn_out: Option<TransactionId>,
 }
 
@@ -84,10 +105,10 @@ impl From<P2PSwapStatus> for AcceptSwapStatusError {
     fn from(value: P2PSwapStatus) -> Self {
         match value {
             P2PSwapStatus::Open => unreachable!(),
-            P2PSwapStatus::Cancelled(s) => AcceptSwapStatusError::OfferCancelled(AcceptSwapOfferCancelled {
+            P2PSwapStatus::Cancelled(s) => AcceptSwapStatusError::SwapCancelled(AcceptSwapSwapCancelled {
                 token0_txn_out: s.token0_txn_out,
             }),
-            P2PSwapStatus::Expired(s) => AcceptSwapStatusError::OfferExpired(AcceptSwapOfferExpired {
+            P2PSwapStatus::Expired(s) => AcceptSwapStatusError::SwapExpired(AcceptSwapSwapExpired {
                 token0_txn_out: s.token0_txn_out,
             }),
             P2PSwapStatus::Reserved(s) => AcceptSwapStatusError::AlreadyReserved(AcceptSwapAlreadyReserved {